Moved tbcsv io ui definitions
