+++ Arma3ObjectBuilder/ui/__init__.py
@@ -2,7 +2,6 @@ from . import import_export_asc
 from . import import_export_mcfg
 from . import import_export_p3d
 from . import import_export_rtm
-from . import import_export_tbcsv
 from . import props_action
 from . import props_material
 from . import props_object_mesh
